File reorganization - Moved seed folder to project root. - Move mail folder out of the /config folder into the /app. - Moved the validation folder into the /config folder. - Wrapped all express features in callable functions.
