+++ app/controllers/feed/entries/types/CommentEntryType.js
@@ -1,5 +1,5 @@
 const { Feed } =  require("../../../../models");
-const { comment: commentVal } = require("../../../../validation");
+const { comment: commentVal } = require("../../../../config/validation");
 const EntryType = require("./EntryType");
 
 class CommentEntryType extends EntryType {
