+++ app/routes/utils/addRequest.js
@@ -4,7 +4,7 @@ const setDefaultError = require("../middleware/setDefaultError");
 const globalParamsValidation = require("../middleware/globalParamsValidation");
 const isAuthenticated = require("../middleware/isAuthenticated");
 const isVerified = require("../middleware/isVerified");
-const ValidationSchema = require("../../validation/ValidationSchema");
+const ValidationSchema = require("../../config/validation/ValidationSchema");
 const PermissionSet = require("../../config/permissions/PermissionSet");
 
 const validationMap = {
