+++ seed/seeds/seedUser.js
@@ -1,4 +1,4 @@
-const { User } = require("../../models");
+const { User } = require("../../app/models");
 
 module.exports = async () => {
